🎨 feat(postProcess): Add post-processing code for 2-LidDrivenCavity-Newtonian.c
This commit adds a new section to the `postProcess/2-casesWithNS.ipynb` notebook that includes post-processing code for the `2-LidDrivenCavity-Newtonian.c` case. The new code generates a plot of the velocity profile along a horizontal line in the lid-driven cavity flow.

The changes include:

- Adding a new markdown cell to introduce the post-processing code for the 2-LidDrivenCavity-Newtonian.c case.
- Adding a new code cell that reads the Basilisk output file, interpolates the x-velocity along a horizontal line, and plots the resulting velocity profile.

These changes will help streamline the post-processing workflow for the lid-driven cavity flow case.
