+++ testCases/2-Poiseuille-flow-Newtonian.c
@@ -1,57 +1,32 @@
 /**
-# Planar Couette flow of Newtonian Fluid
-
-This code simulates a planar Couette flow with a Newtonian fluid.
-It uses a constant viscosity model, which is a simplified version of
-the generalized non-Newtonian model in 2-Couette-flow-Bingham.c.
-
-## Mathematical Formulations
-
-For Newtonian fluids, the stress-strain rate relationship is linear:
-$$ \tau = 2\mu_0 D_{ij} $$
-
-Where $D_{ij}$ is the deformation tensor:
-$D_{ij}=(u_{i,j}+u_{j,i})/2$
-
-## Exact solution in the proposed case
-
-For a Newtonian fluid with constant viscosity $\mu_0$ and a constant 
-pressure gradient $-\frac{dp}{dx} = 1$, the velocity profile is parabolic:
-
-$$ u(y) = \frac{1}{2\mu_0}y(h-y) $$
-
-Where $h$ is the height of the domain (1.0 in this case).
-
+#  Planar Couette flow of Newtonian Fluid
 # Code
 */
-
 #include "navier-stokes/centered.h"
 
 // Global parameters
 char file_name[80];
-double mu_0 = 1.0;        // Constant viscosity (Newtonian)
-int max_iter = 1e4;       // Maximum iterations
-#define DT_MAX (1e-3)     // Maximum timestep
+double tau_y = 0.0;        // Yield stress
+double mu_0 = 1.0;         // Base viscosity
+int max_iter = 1e4;        // Maximum iterations
+#define DT_MAX (1e-3)      // Maximum timestep
 
 int main(int argc, char const *argv[])
 {
-  if (argc < 2) {
-    fprintf(stderr, "Usage: %s filename\n", argv[0]);
-    return 1;
-  }
   
-  sprintf(file_name, "%s", argv[1]);
+  sprintf(file_name, "results");
   
   // Initialize grid and domain
-  init_grid(1<<6);
+  init_grid(1<<8);
   L0 = 1.0;
-  origin(0.0, 0.0);
+  origin(-0.5, -0.5);
   DT = DT_MAX;
   stokes = true;
   TOLERANCE = 1e-5;
 
 /**
- For Newtonian fluid, we only need to set the constant viscosity $\mu_0$
+ Values of yield stress, viscosity, and coefficient.
+ - Newtonian: $\mu_0 = 1.0$
 */
   // Set boundary conditions
   // Right-left boundaries are periodic
@@ -64,7 +39,6 @@ int main(int argc, char const *argv[])
   
   // No slip at the bottom
   u.n[bottom] = dirichlet(0);
-  uf.n[bottom] = dirichlet(0);
   u.t[bottom] = dirichlet(0);
   
   // Pressure conditions are neumann 0.0
@@ -78,14 +52,15 @@ int main(int argc, char const *argv[])
 
 // un is used to search for a stationary solution
 scalar un[];
+// muv will be used as the face vector for viscosity
+face vector muv[];
 
 /**
 ## Initialization event
 */
 event init(t = 0) {
-  // Set constant Newtonian viscosity
-  const scalar mu_const[] = mu_0;
-  mu = mu_const;
+  // Set Non-Newtonian viscosity
+  mu = muv;
   
   /**
    Pressure gradient `mdpdx`
@@ -123,22 +98,11 @@ event logfile(i += 500; i <= max_iter) {
 }
 
 /**
-## Compare with exact solution
+## Calculating viscosity for generalized Newtonian fluid
 */
-event end(t = end) {
-  FILE *fp = fopen("profile.dat", "w");
-  fprintf(fp, "# y u_numerical u_exact error\n");
-  
-  // Sample the velocity profile along the x-midpoint
-  foreach_col(0) {
-    // Calculate exact solution: u(y) = (1/2μ) * y * (h-y)
-    double y = y;
-    double u_exact = (1.0/(2.0*mu_0)) * y * (L0-y);
-    double error = fabs((u.x[] - u_exact)/u_exact);
-    
-    fprintf(fp, "%g %g %g %g\n", y, u.x[], u_exact, error);
+event properties(i++) {
+  foreach_face() {
+    // Apply viscosity at face
+    muv.x[] = fm.x[] * mu_0;
   }
-  
-  fclose(fp);
-  fprintf(ferr, "Profile data written to 'profile.dat'\n");
 }
